+++ src/main/java/vn/thodev/model/inventory/Docket.java
@@ -46,7 +46,7 @@ public class Docket extends AbstractEntity<Long> {
     @Column(name = "note")
     private String note;
 
-    @Column(name = "status", nullable = false, columnDefinition = "TINYINT")
+    @Column(name = "status", nullable = false, columnDefinition = "SMALLINT")
     private Integer status;
 
 }
